+++ growth/feature-flag-client/src/lib.ts
@@ -1,32 +1,31 @@
-import { FlagShape, Flags, CustomAttributes, Reason, ErrorKind } from './types';
+import { FlagShape, CustomAttributes, FlagType, Flags } from './types';
 
-export const isType = (value: any, type: string): boolean => {
-  return value !== null && typeof value === type;
+export const getFlagType = (value: any): FlagType => {
+  switch (typeof value) {
+    case 'boolean':
+      return FlagType.BOOLEAN;
+    case 'string':
+      return FlagType.STRING;
+    case 'object':
+      if (value === null) {
+        return FlagType.UNKNOWN;
+      }
+      return FlagType.JSON;
+    default:
+      return FlagType.UNKNOWN;
+  }
 };
 
-export const isObject = (value: any) => isType(value, 'object');
-export const isBoolean = (value: any) => isType(value, 'boolean');
-export const isString = (value: any) => isType(value, 'string');
-
-export const isFlagWithEvaluationDetails = (flag: FlagShape): boolean => {
-  return isObject(flag) && 'value' in flag && 'explanation' in flag;
-};
-
-export const isSimpleFlag = (flag: FlagShape): boolean => {
-  return isObject(flag) && 'value' in flag && !('explanation' in flag);
-};
-
-export const isOneOf = (value: string, list: string[]): boolean =>
-  list.indexOf(value) > -1;
+export const isObject = (value: any) => typeof value === 'object';
 
 export const enforceAttributes = (
   obj: any,
   attributes: string[],
   identifier?: string,
 ) => {
-  const title = identifier ? `${identifier}: ` : '';
   attributes.forEach((attribute: string) => {
-    if (!obj.hasOwnProperty(attribute) && obj[attribute] !== null) {
+    if (!(attribute in obj)) {
+      const title = identifier ? `${identifier}: ` : '';
       throw new Error(`${title}Missing ${attribute}`);
     }
   });
@@ -54,7 +53,7 @@ export const checkForReservedAttributes = (
 };
 
 const validateFlag: any = (flagKey: string, flag: FlagShape) => {
-  if (isSimpleFlag(flag) || isFlagWithEvaluationDetails(flag)) {
+  if (isObject(flag) && 'value' in flag) {
     return true;
   }
 
@@ -68,28 +67,4 @@ const validateFlag: any = (flagKey: string, flag: FlagShape) => {
 
 export const validateFlags = (flags: Flags) => {
   Object.keys(flags).forEach(key => validateFlag(key, flags[key]));
-};
-
-const setReasonAndKindErrors = (flag: FlagShape): FlagShape['explanation'] => {
-  let explanation = {
-    kind: 'ERROR' as Reason,
-    errorKind: 'WRONG_TYPE' as ErrorKind,
-  };
-  if (flag.explanation) {
-    explanation = {
-      ...flag.explanation,
-      ...explanation,
-    };
-  }
-  return explanation;
-};
-
-export const validateFlagExplanation = (
-  flag: FlagShape,
-  defaultValue: boolean | string | object,
-): FlagShape['explanation'] => {
-  if (typeof flag.value !== typeof defaultValue) {
-    return setReasonAndKindErrors(flag);
-  }
-  return flag.explanation;
 };