+++ growth/feature-flag-client/src/types.ts
@@ -7,19 +7,42 @@ export type Reason =
   | 'SIMPLE_EVAL'
   | 'ERROR';
 
-export type ErrorKind = 'WRONG_TYPE' | 'FLAG_NOT_FOUND';
+export type ErrorKind = 'WRONG_TYPE' | 'FLAG_NOT_FOUND' | 'VALIDATION_ERROR';
 
 export type RuleId = string;
 
-export type FlagShape = {
-  value: boolean | string | object;
-  explanation?: {
-    kind: Reason;
-    ruleId?: RuleId;
-    ruleIndex?: number;
-    errorKind?: ErrorKind;
-  };
+export type FlagValue = boolean | string | object;
+
+export type FlagExplanation = {
+  kind: Reason;
+  ruleId?: RuleId;
+  ruleIndex?: number;
+  errorKind?: ErrorKind;
 };
+
+export type FlagShape = {
+  value: FlagValue;
+  explanation?: FlagExplanation;
+};
+
+export type TriggeredExposureHandler = (
+  flagKey: string,
+  flag: FlagShape,
+  exposureData?: CustomAttributes,
+) => void;
+
+export type AutomaticExposureHandler = (
+  flagKey: string,
+  value: string | boolean | object,
+  flagExplanation?: FlagShape['explanation'],
+) => void;
+
+export enum FlagType {
+  BOOLEAN,
+  JSON,
+  STRING,
+  UNKNOWN,
+}
 
 export type Flags = {
   [flagName: string]: FlagShape;
@@ -29,7 +52,7 @@ export type ReservedAttributes = {
   flagKey: string;
   reason: Reason;
   ruleId?: string;
-  value: boolean | string | object;
+  value: FlagValue;
   errorKind?: ErrorKind;
 };
 
@@ -48,28 +71,28 @@ export type ExposureEvent = {
   highPriority?: boolean;
 };
 
-export interface FlagConstructor {
-  new (
-    flagKey: string,
-    flag: FlagShape,
-    trackExposure: (flagKey: string, flag: FlagShape) => void,
-  ): Flag;
-}
-export interface Flag {
+export interface FlagWrapper {
+  evaluationCount: number;
   getBooleanValue(options: {
     default: boolean;
     shouldTrackExposureEvent?: boolean;
-    data: CustomAttributes;
+    exposureData?: CustomAttributes;
   }): boolean;
 
   getVariantValue(options: {
     default: string;
     oneOf: string[];
     shouldTrackExposureEvent?: boolean;
-    data: CustomAttributes;
+    exposureData?: CustomAttributes;
   }): string;
 
   getJSONValue(): object;
+
+  getRawValue(options: {
+    default: FlagValue;
+    shouldTrackExposureEvent?: boolean;
+    exposureData?: CustomAttributes;
+  }): FlagValue;
 }
 
 export type AnalyticsHandler = (event: ExposureEvent) => void;
@@ -77,3 +100,9 @@ export type AnalyticsHandler = (event: ExposureEvent) => void;
 export interface AutomaticAnalyticsHandler {
   sendOperationalEvent(event: ExposureEvent): void;
 }
+
+export type FlagStats = {
+  [flagKey: string]: {
+    evaluationCount: number;
+  };
+};